Optionally track more stats
**WIP**

There are some additional stats that could be tracked:

```
averagerating
ratingcount
trendingdaily
trendingmonthly
trendingweekly
weightedRating
```

This is a POC of how it could be done.  It defaults to the existing stats. If you uncomment lines 22-23, it will then also track `trendingweekly` & `trendingdaily`.

If there is an existing ``` `${extensionName}-stats-${isoDateString}.csv` ``` file with matching CSV headers, it will write to it. If the headers are wrong, it will append a number to end, incrementing this number as needed: ``` `${extensionName}-stats-${isoDateString}_${suffix++}.csv` ```

Some of the changes (the most glaring being the switch from `require` to `import`) are due to the use of a top-level `await`.

---

While throwing this together, I ran across a bug introduced in https://github.com/RandomFractals/vscode-marketplace-ext-stats/pull/5 -- the CSV headers were only changed from `DateTime, Installs, Downloads, Version` to `DateTime, Installs, Updates, Downloads, Version` in the `console.log`, they should've also been changed on [line 65](https://github.com/RandomFractals/vscode-marketplace-ext-stats/blob/0224c96/tools/vscode-marketplace-stats.js#L65) where they are written to the `.csv` file (as well as in a few comments).